# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and Authorization API",
        "title": "Auth API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/pdf/pdf-to-word": {
            "post": {
                "description": "PDF faylni Word formatiga o‘tkazish",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF to Word"
                ],
                "summary": "Convert PDF to Word",
                "parameters": [
                    {
                        "description": "PDF to Word request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PDFToWordRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/pdf-to-word/{id}": {
            "get": {
                "description": "PDF to Word ish holatini ko‘rish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF to Word"
                ],
                "summary": "Get PDF to Word Job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PDFToWordJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/ppt-to-pdf": {
            "post": {
                "description": "PowerPoint faylni PDF formatga o‘tkazish",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PowerPoint to PDF"
                ],
                "summary": "Convert PowerPoint to PDF",
                "parameters": [
                    {
                        "description": "PowerPoint to PDF request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PowerPointToPDFRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "job_id qaytadi",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/ppt-to-pdf/{id}": {
            "get": {
                "description": "Konvertatsiya holatini olish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PowerPoint to PDF"
                ],
                "summary": "Get PowerPoint→PDF Job Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PowerPointToPDFJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/protect": {
            "post": {
                "description": "PDF fayliga parol qo‘shadi (foydalanuvchi tomonidan berilgan parol asosida)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Security"
                ],
                "summary": "Protect PDF with password",
                "parameters": [
                    {
                        "description": "Protect request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProtectPDFRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/protect/{id}": {
            "get": {
                "description": "Parollangan PDF faylga oid ish holatini qaytaradi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Security"
                ],
                "summary": "Get protected PDF job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProtectPDFJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/removepage": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-remove"
                ],
                "summary": "Create remove pages job",
                "parameters": [
                    {
                        "description": "remove pages job",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RemovePagesRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/removepage/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-remove"
                ],
                "summary": "Get remove pages job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "remove job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RemoveJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.PDFToWordJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.PDFToWordRequest": {
            "type": "object",
            "required": [
                "input_file_id"
            ],
            "properties": {
                "input_file_id": {
                    "type": "string"
                }
            }
        },
        "models.PowerPointToPDFJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.PowerPointToPDFRequest": {
            "type": "object",
            "required": [
                "input_file_id"
            ],
            "properties": {
                "input_file_id": {
                    "type": "string"
                }
            }
        },
        "models.ProtectPDFJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "description": "← *pointer (✔ to‘g‘ri)",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.ProtectPDFRequest": {
            "type": "object",
            "required": [
                "input_file_id",
                "password"
            ],
            "properties": {
                "input_file_id": {
                    "description": "PDF fayl ID",
                    "type": "string"
                },
                "password": {
                    "description": "Qo‘yiladigan parol",
                    "type": "string"
                }
            }
        },
...
        "models.RemoveJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "description": "NULL bo‘lishi mumkin",
                    "type": "string"
                },
                "pages_to_remove": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "description": "NULL bo‘lishi mumkin",
                    "type": "string"
                }
            }
        },
        "models.RemovePagesRequest": {
            "type": "object",
            "properties": {
                "input_file_id": {
                    "type": "string"
                },
                "pages_to_remove": {
                    "description": "e.g. \"1,4-6\"",
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}